exporter/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
click==8.3.1
colorama==0.4.6
fastapi==0.129.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.1
prometheus_client==0.24.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
yarl==1.22.0
